Fixed a few warnings to please the compiler

+++ AppleIntelWiFiMVM/linux/iwl-fw.h
@@ -129,7 +129,7 @@ fw_has_capa(const struct iwl_ucode_capabilities *capabilities,
 /* one for each uCode image (inst/data, init/runtime/wowlan) */
 struct fw_desc {
 	const void *data;	/* vmalloc'ed data */
-	u32 len;		/* size in bytes */
+	size_t len;		/* size in bytes */
 	u32 offset;		/* offset in the device */
 };
